requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.14
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
python-dotenv==1.0.1
psycopg2==2.9.9
django-crispy-forms==2.3
crispy-bootstrap4==2024.1
Pillow==10.4.0





